Workaround for modified icons in the file chooser being pixelated
Comment also added to the source (since this workaround is not optimal, and I would like to find the correct solution):

If we use `self.size` for the off-screen image buffer then we get a scaled up 16×16 image in the file chooser (⌘T).

My theory is that when drawing the image, the image size is 16×16 but the graphics context uses a transformation so that each point is multiple pixels, which our off-screen image does not replicate.

I do not know how to get the actual “pixel size” of the destination, so using the largest image is a workaround and knowing the actual size could give a better result.